combined the admin single order store and the other single order store
